takehometest/requirements.txt
boto3==1.21.0
psycopg2-binary==2.9.3
pytest==6.2.4
